+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.2.0...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- Prevents `get_packages` from returning dot-separated subdirectories
 
 [3.2.0](https://github.com/jshwi/pyaud/releases/tag/v3.2.0) - 2021-08-12
 ------------------------------------------------------------------------